Update bin default to include entities
This extends the prior fix to include entity functionality to the
default flow to the commandline bin file:

Added :entities to the operations array. This will result in special
characters being replaced with their corresponding entities along with
all the other filters in the default mode when using the commandline,
as promised.

Apologies for overlooking this in the previous PR!